+++ src/core/json/stream/streamers/index.ts
@@ -1,0 +1,18 @@
+/*!
+ * V4Fire Core
+ * https://github.com/V4Fire/Core
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Core/blob/master/LICENSE
+ */
+
+/**
+ * [[include:core/json/stream/streamers/README.md]]
+ * @packageDocumentation
+ */
+
+export { default as ArrayStreamer } from 'core/json/stream/streamers/array';
+export { default as ObjectStreamer } from 'core/json/stream/streamers/object';
+export { default as AbstractStreamer } from 'core/json/stream/streamers/interface';
+
+export * from 'core/json/stream/streamers/interface';