+++ src/core/object/proxy-readonly/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.75.0 (2022-02-22)
+
+#### :rocket: New Feature
+
+* Added a common function to create a read-only view
+
 ## v3.73.0 (2021-12-14)
 
 #### :rocket: New Feature